+++ lib/eve_online/esi/models/region.rb
@@ -6,15 +6,10 @@ module EveOnline
       class Region < Base
         def as_json
           {
-            constellations: constellations, # TODO: rename constellations -> constellation_ids and remove from as_json
             description: description,
             name: name,
             region_id: region_id,
           }
-        end
-
-        def constellations
-          options["constellations"]
         end
 
         def description
@@ -27,6 +22,10 @@ module EveOnline
 
         def region_id
           options["region_id"]
+        end
+
+        def constellation_ids
+          options["constellations"]
         end
       end
     end