+++ spec/eve_online/esi/models/region_spec.rb
@@ -16,8 +16,6 @@ describe EveOnline::ESI::Models::Region do
   describe "#as_json" do
     let(:region) { described_class.new(options) }
 
-    before { expect(region).to receive(:constellations).and_return([20_000_001]) }
-
     before { expect(region).to receive(:description).and_return("The Derelik region...") }
 
     before { expect(region).to receive(:name).and_return("Derelik") }
@@ -26,19 +24,11 @@ describe EveOnline::ESI::Models::Region do
 
     subject { region.as_json }
 
-    its([:constellations]) { should eq([20_000_001]) }
-
     its([:description]) { should eq("The Derelik region...") }
 
     its([:name]) { should eq("Derelik") }
 
     its([:region_id]) { should eq(10_000_001) }
-  end
-
-  describe "#constellations" do
-    before { expect(options).to receive(:[]).with("constellations") }
-
-    specify { expect { subject.constellations }.not_to raise_error }
   end
 
   describe "#description" do
@@ -57,5 +47,11 @@ describe EveOnline::ESI::Models::Region do
     before { expect(options).to receive(:[]).with("region_id") }
 
     specify { expect { subject.region_id }.not_to raise_error }
+  end
+
+  describe "#constellation_ids" do
+    before { expect(options).to receive(:[]).with("constellations") }
+
+    specify { expect { subject.constellation_ids }.not_to raise_error }
   end
 end