+++ spec/eve_online/esi/universe_region_spec.rb
@@ -68,16 +68,6 @@ describe EveOnline::ESI::UniverseRegion do
     specify { expect { subject.as_json }.not_to raise_error }
   end
 
-  describe "#constellations" do
-    let(:model) { instance_double(EveOnline::ESI::Models::Region) }
-
-    before { subject.instance_variable_set(:@model, model) }
-
-    before { expect(model).to receive(:constellations) }
-
-    specify { expect { subject.constellations }.not_to raise_error }
-  end
-
   describe "#description" do
     let(:model) { instance_double(EveOnline::ESI::Models::Region) }
 
@@ -106,6 +96,16 @@ describe EveOnline::ESI::UniverseRegion do
     before { expect(model).to receive(:region_id) }
 
     specify { expect { subject.region_id }.not_to raise_error }
+  end
+
+  describe "#constellation_ids" do
+    let(:model) { instance_double(EveOnline::ESI::Models::Region) }
+
+    before { subject.instance_variable_set(:@model, model) }
+
+    before { expect(model).to receive(:constellation_ids) }
+
+    specify { expect { subject.constellation_ids }.not_to raise_error }
   end
 
   describe "#scope" do