+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## master
+## v0.31.0
 
 * Add `EveOnline::ESI::Models::MailHeader`
 * Add `EveOnline::ESI::Models::MailRecipient`